refactor granular synthesizer to require fewer setup calls

+++ web/src/audio/defaults.rs
@@ -2,8 +2,8 @@
 /// is updated with the true sample rate from the audio context.
 pub const FALLBACK_SAMPLE_RATE: u32 = 44100;
 
-pub const MAX_NUM_CHANNELS: usize = 250;
+pub const MAX_NUM_CHANNELS: u32 = 250;
 
-pub const GRAIN_LEN_MIN_IN_MS: usize = 10;
+pub const GRAIN_LEN_MIN_IN_MS: u32 = 10;
 
-pub const GRAIN_LEN_MAX_IN_MS: usize = 1000;
+pub const GRAIN_LEN_MAX_IN_MS: u32 = 1000;
